# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vaishnavi Group Help Desk API
  version: 1.0.0
  description: >
    REST API for the Help Desk / Ticketing system. All responses are JSON and
    follow the envelope `{ "success": boolean, "data"?: ..., "message"?: string }`.
    Paginated list endpoints additionally return `meta: { page, limit, total, totalPages }`.
    Authenticate with a Bearer JWT access token obtained from `/auth/login`.
servers:
  - url: http://localhost:5001/api/v1
    description: Local development
  - url: https://YOUR_API_DOMAIN/api/v1
    description: Production (replace with deployed API host)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /knowledge-base:
    get:
      tags: [KnowledgeBase]
      summary: List articles (published only for end users)
      parameters:
        - { in: query, name: search, schema: { type: string } }
        - { in: query, name: category, schema: { type: string } }
      responses: { '200': { description: Articles } }
    post:
      tags: [KnowledgeBase]
      summary: Create an article (admin/manager)
      responses: { '201': { description: Created } }
  /knowledge-base/{slug}:
    get:
      tags: [KnowledgeBase]
      summary: Get an article by slug (with attachments; increments views)
      parameters: [{ in: path, name: slug, required: true, schema: { type: string } }]
      responses: { '200': { description: Article } }
  /knowledge-base/{id}:
    patch:
      tags: [KnowledgeBase]
      summary: Update an article (admin/manager)
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses: { '200': { description: Updated } }
    delete:
      tags: [KnowledgeBase]
      summary: Delete an article (admin/manager)
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses: { '200': { description: Deleted } }
  /knowledge-base/{id}/attachments:
    post:
      tags: [KnowledgeBase]
      summary: Upload an attachment to an article (multipart, field "file") (admin/manager)
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      requestBody:
        content:
          multipart/form-data:
            schema: { type: object, properties: { file: { type: string, format: binary } } }
      responses: { '201': { description: Uploaded } }
  /knowledge-base/attachments/{attId}:
    delete:
      tags: [KnowledgeBase]
      summary: Delete an article attachment (admin/manager)
      parameters: [{ in: path, name: attId, required: true, schema: { type: string } }]
      responses: { '200': { description: Deleted } }
...